CDC-15 Refactor: supabase libs
