+++ apps/native/app/(main)/(tabs)/(home)/sales.tsx
@@ -1,10 +1,10 @@
-import { CategorySectors } from '@cococom/supabase/libs';
 import { BottomSheetModal } from '@gorhom/bottom-sheet';
 import { router, Stack, useLocalSearchParams } from 'expo-router';
 import React, { ComponentType, useCallback, useMemo, useRef, useState } from 'react';
 import { useWindowDimensions, View } from 'react-native';
 import { Route, SceneMap, TabBar, TabView, TabViewProps } from 'react-native-tab-view';
 import { createStyleSheet, useStyles } from 'react-native-unistyles';
+import { CategorySectors } from '../../../../../../packages/supabase/dist/lib/lib';
 
 import Chip from '@/components/core/chip';
 import SortBottomSheet from '@/components/custom/bottom-sheet/sort';
