+++ apps/native/libs/sort/ranking.ts
@@ -1,5 +1,5 @@
-import { SortOptionDirection } from '@cococom/supabase/libs';
 import { AlltimeRankingResultItem } from '@cococom/supabase/types';
+import { SortOptionDirection } from '../../../../packages/supabase/dist/lib/lib';
 
 import { DiscountSortOption } from './discount';
