+++ packages/eslint-config/node.js
@@ -6,11 +6,11 @@ module.exports = {
   parser: "@typescript-eslint/parser",
   extends: [
     "plugin:@typescript-eslint/recommended",
-    "prettier",
     "eslint:recommended",
+    "prettier",
     "turbo",
   ],
-  plugins: ["import", "prettier"],
+  plugins: ["prettier", "import"],
   parserOptions: {
     ecmaVersion: 2020,
     sourceType: "module",
@@ -35,8 +35,7 @@ module.exports = {
           "type",
           "builtin",
           "external",
-          "internal",
-          ["parent", "sibling"],
+          ["internal", "parent", "sibling"],
           "index",
           "unknown",
         ],
@@ -46,10 +45,15 @@ module.exports = {
             group: "external",
             position: "before",
           },
+          {
+            pattern: "@/**",
+            group: "internal",
+            position: "after",
+          },
         ],
         "newlines-between": "always",
         alphabetize: {
-          order: "asc",
+          order: "ignore",
           caseInsensitive: true,
         },
       },
