+++ packages/supabase/src/merged-types.ts
@@ -170,3 +170,16 @@ export type Database = MergeDeep<
     };
   }
 >;
+
+export type InsertDiscount = Database['public']['Tables']['discounts']['Insert'];
+export type InsertItem = Database['public']['Tables']['items']['Insert'];
+export type InsertCategory = Database['public']['Tables']['categories']['Insert'];
+export type InsertWishlist = Omit<
+  Database['public']['Tables']['wishlists']['Insert'],
+  'wishlist_hash'
+>;
+export type InsertHistory = Database['public']['Tables']['histories']['Insert'];
+export type InsertMemo = Database['public']['Tables']['memos']['Insert'];
+export type InsertComment = Database['public']['Tables']['comments']['Insert'];
+export type CategorySectors = Database['public']['Enums']['CategorySectors'];
+export type SortOptionDirection = 'ASC' | 'DESC';
